pyproject.toml
[project]
name = "mcp-a2a-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.16",
    "google-adk>=1.18.0",
    "mcp[cli]>=1.21.1",
]
